fix: re-apply u64 overflow fix (regressed in #2 squash) + research doc
The #2 squash-merge dropped simd_gemm.rs changes in a merge conflict —
main still has `if m * n * k < THRESHOLD` which overflows 32-bit usize
to 0 at (2048, 2048, 2048), silently routing to single-threaded. This
is the same bug fixed in the PR's 0320aad.

Also: the 'cache associativity aliasing' explanation in
GEMM-OPTIMIZATION-NOTES.md was the initial misdiagnosis (preserved for
posterity, marked as such). The 2040/2048/2056 pattern is exactly what
overflow-masquerading-as-something-else looks like.

Adds docs/GEMM-NEXT-STEPS.md — deep research synthesis covering:

  WHY WE BEAT XNNPACK (from disassembling their WASM):
    * tf.js 4.22 ships NO relaxed-simd build — they use separate
      f32x4.mul + f32x4.add; we use f32x4.relaxed_madd → native FMA.
      That's 2× fewer arith instructions, the core of our lead.
    * They use MR=5 (10 acc), we use MR=6 (12 acc) — 20% more compute
      per B-load. We're at V8's 15-XMM ceiling exactly.
    * Their x86 config (gemm-config.c:942) never tried 6×8 loadsplat —
      we found a hole in their search space.

  OUR CODEGEN ISSUES (from disassembling rumpy_wasm_bg.wasm):
    * LLVM spilled the 6 A-row pointers, rechains them every K iter
      (14 extra WASM instructions). Fix: black_box or bump-in-loop.
    * No K-unroll (11 instrs loop overhead / K). Fix: manual unroll-2.
    * v128.load offset=N not folded (3-18 wasted instrs). LLVM bug.
    * Expected gain from all three: ~5-10%.

  TIER 1 NEXT STEPS (high ROI, known to work):
    * GEMMFIP fused first-pass packing (ICS 2023) — eliminates the
      'when to pack A' heuristic. 5-15%.
    * K-splitting for skinny shapes (attention score·V, K-reduction).
      oneDNN gate: (m/64 + n/64) < nthreads AND k ≥ 2048. Up to 3-4×
      where we currently get 1-core.
    * Per-browser kernel selection — Firefox prefers deep K-unroll.

  TIER 2 EXPERIMENTS:
    * Hilbert-curve block scheduling (Intel, Jan 2026) — 30 LoC, 1.4×
      over oneDNN on many-core. Shape-oblivious (good for WASM).
    * tinyBLAS k-vectorized kernel for small matrices — nobody ported
      to WASM yet. Zero packing.
    * LLVM RegStackify barriers (llvm#98631 open, affects us).

  DON'T BOTHER (with receipts):
    * Strassen — zero production ML libraries ship it, 3-8% at n=2048.
    * Prefetch — rejected by its own WASM proposal author, -2 to -8%.
    * NT stores — neither BLIS nor oneDNN uses them in GEMM.
    * Runtime low-rank detection — searched all major libs, none do it.

  WATCH: when tf.js ships relaxed-simd, our lead roughly halves.

+++ crates/rumpy-cpu/src/simd_gemm.rs
@@ -2219,7 +2219,9 @@ pub fn matmul_parallel_f32(a: &[f32], b: &[f32], m: usize, n: usize, k: usize) -
     use rayon::prelude::*;
 
     // For small matrices, single-threaded is faster (no thread overhead)
-    if m * n * k < 64 * 64 * 64 {
+    // CRITICAL: WASM usize is 32-bit. m*n*k at 2048³ = 8.6e9 overflows to 0
+    // → silent single-threaded fallback. See docs/GEMM-OPTIMIZATION-NOTES.md.
+    if (m as u64) * (n as u64) * (k as u64) < (64u64 * 64 * 64) {
         return matmul_dispatch_f32(a, b, m, n, k);
     }
 
@@ -2268,7 +2270,9 @@ pub fn matmul_parallel_f32_v2(a: &[f32], b: &[f32], m: usize, n: usize, k: usize
     use rayon::prelude::*;
 
     // For small matrices, single-threaded is faster (no thread overhead)
-    if m * n * k < 64 * 64 * 64 {
+    // CRITICAL: WASM usize is 32-bit. m*n*k at 2048³ = 8.6e9 overflows to 0
+    // → silent single-threaded fallback. See docs/GEMM-OPTIMIZATION-NOTES.md.
+    if (m as u64) * (n as u64) * (k as u64) < (64u64 * 64 * 64) {
         return matmul_dispatch_f32(a, b, m, n, k);
     }
 
@@ -2318,7 +2322,9 @@ pub fn matmul_pthreadpool_f32(a: &[f32], b: &[f32], m: usize, n: usize, k: usize
     use pthreadpool_rs::ThreadPool;
 
     // For small matrices, single-threaded is faster (no thread overhead)
-    if m * n * k < 64 * 64 * 64 {
+    // CRITICAL: WASM usize is 32-bit. m*n*k at 2048³ = 8.6e9 overflows to 0
+    // → silent single-threaded fallback. See docs/GEMM-OPTIMIZATION-NOTES.md.
+    if (m as u64) * (n as u64) * (k as u64) < (64u64 * 64 * 64) {
         return matmul_dispatch_f32(a, b, m, n, k);
     }
 
@@ -2380,7 +2386,9 @@ pub fn matmul_pthreadpool_f32_with_pool(
     k: usize,
 ) -> Vec<f32> {
     // For small matrices, single-threaded is faster (no thread overhead)
-    if m * n * k < 64 * 64 * 64 {
+    // CRITICAL: WASM usize is 32-bit. m*n*k at 2048³ = 8.6e9 overflows to 0
+    // → silent single-threaded fallback. See docs/GEMM-OPTIMIZATION-NOTES.md.
+    if (m as u64) * (n as u64) * (k as u64) < (64u64 * 64 * 64) {
         return matmul_dispatch_f32(a, b, m, n, k);
     }
 
@@ -2875,8 +2883,13 @@ pub fn matmul_optimized_f32(a: &[f32], b: &[f32], m: usize, n: usize, k: usize) 
 pub fn matmul_optimized_f32_parallel(a: &[f32], b: &[f32], m: usize, n: usize, k: usize) -> Vec<f32> {
     use rayon::prelude::*;
 
-    // For small matrices, single-threaded is faster
-    if m * n * k < 128 * 128 * 128 {
+    // For small matrices, single-threaded is faster.
+    //
+    // CRITICAL: WASM usize is 32-bit. At (2048, 2048, 2048), m*n*k = 8.6e9
+    // overflows to EXACTLY 0 → this check passes → silent single-threaded
+    // fallback → "1.00× speedup" that looked like cache aliasing for 6 hours.
+    // Always use u64 for flop estimates on wasm32.
+    if (m as u64) * (n as u64) * (k as u64) < (128u64 * 128 * 128) {
         return matmul_optimized_f32(a, b, m, n, k);
     }
